fix: fixing id types

+++ src/main/resources/db/migration/V4__create-table-movimentacoes.sql
@@ -1,6 +1,6 @@
 CREATE TABLE movimentacoes (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    produto_id INT NOT NULL,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    produto_id BIGINT NOT NULL,
     tipo VARCHAR(50) NOT NULL,
     quantidade INT NOT NULL,
     data DATETIME DEFAULT CURRENT_TIMESTAMP,
